refactor(cli): move registry cache removal into SkillManager
Commands stay orchestration-only: validate, guard, delegate, render. The containment check and fs removal now live beside the rest of the cache-path logic in SkillManager, with unit tests for the happy path and escape refusal.
